+++ implementation/Scala/src/main/scala/synapses/Library.scala
@@ -36,6 +36,30 @@ object Library {
       Network.init(layerSizes, activationF)(weightInitF)
     }
 
+    private def throwIfInputNotMatch(network: NeuralNetwork,
+                                     inputValues: List[Double])
+    : Unit = {
+      val numOfInputVals = inputValues.length
+      val inputLayerSize = network.head.head.weights.length - 1
+      if (numOfInputVals != inputLayerSize)
+        throw new Exception(
+          s"the number of input values ($numOfInputVals) " +
+            s"does not match the size of the input layer ($inputLayerSize)"
+        )
+    }
+
+    private def throwIfExpectedNotMatch(network: NeuralNetwork,
+                                        expectedOutput: List[Double])
+    : Unit = {
+      val numOfExpectedVals = expectedOutput.length
+      val outputLayerSize = network.last.length
+      if (numOfExpectedVals != outputLayerSize)
+        throw new Exception(
+          s"the number of expected values ($numOfExpectedVals)" +
+            s" does not match the size of the output layer ($outputLayerSize)"
+        )
+    }
+
     def init(layers: List[Int]): NeuralNetwork =
       seedInit(None, layers)
 
@@ -53,6 +77,7 @@ object Library {
     def prediction(network: NeuralNetwork,
                    inputValues: List[Double])
     : List[Double] = {
+      throwIfInputNotMatch(network, inputValues)
       val input = inputValues.to(LazyList)
       Network
         .output(input)(network)
@@ -64,6 +89,8 @@ object Library {
                inputValues: List[Double],
                expectedOutput: List[Double])
     : List[Double] = {
+      throwIfInputNotMatch(network, inputValues)
+      throwIfExpectedNotMatch(network, expectedOutput)
       val input = inputValues.to(LazyList)
       val expected = expectedOutput.to(LazyList)
       Network
@@ -76,6 +103,8 @@ object Library {
             inputValues: List[Double],
             expectedOutput: List[Double])
     : NeuralNetwork = {
+      throwIfInputNotMatch(network, inputValues)
+      throwIfExpectedNotMatch(network, expectedOutput)
       val input = inputValues.to(LazyList)
       val expected = expectedOutput.to(LazyList)
       Network.fit(learningRate, input, expected)(network)